LPS-61577 Change package and metadata
