+++ modules/portal/portal-metatype-annotations/bnd.bnd
@@ -1,7 +1,7 @@
-Bundle-Name: Liferay Configuration Admin Annotations
-Bundle-SymbolicName: com.liferay.configuration.admin.annotations
+Bundle-Name: Liferay Metatype Definitons Annotations
+Bundle-SymbolicName: com.liferay.portal.metatype.definitions.annotations
 Bundle-Version: 1.0.0
 Include-Resource: classes
 Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Configuration Framework
+Liferay-Releng-Module-Group-Title: Metatype
 -include: ../../../../marketplace/platform/bnd.bnd
