+++ modules/portal/portal-metatype-annotations/src/main/java/com/liferay/portal/metatype/definitions/annotations/AnnotationsExtendedObjectClassDefinition.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.configuration.admin.annotations;
+package com.liferay.portal.metatype.definitions.annotations;
 
-import com.liferay.configuration.admin.ExtendedAttributeDefinition;
 import com.liferay.portal.metatype.annotations.ExtendedObjectClassDefinition;
+import com.liferay.portal.metatype.definitions.ExtendedAttributeDefinition;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -34,7 +34,7 @@ import org.osgi.service.metatype.ObjectClassDefinition;
  * @author Iván Zaera
  */
 public class AnnotationsExtendedObjectClassDefinition
-	implements com.liferay.configuration.admin.ExtendedObjectClassDefinition {
+	implements com.liferay.portal.metatype.definitions.ExtendedObjectClassDefinition {
 
 	public AnnotationsExtendedObjectClassDefinition(
 		Bundle bundle, ObjectClassDefinition objectClassDefinition) {
