+++ modules/portal/portal-metatype-equinox/src/main/java/com/liferay/portal/metatype/definitions/equinox/EquinoxExtendedMetaTypeService.java
@@ -12,27 +12,25 @@
  * details.
  */
 
-package com.liferay.configuration.admin.annotations;
+package com.liferay.portal.metatype.definitions.equinox;
 
 import com.liferay.portal.metatype.definitions.ExtendedMetaTypeInformation;
 import com.liferay.portal.metatype.definitions.ExtendedMetaTypeService;
 
 import org.osgi.framework.Bundle;
-import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.metatype.MetaTypeService;
 
 /**
  * @author Iván Zaera
  */
-@Component(service = ExtendedMetaTypeService.class)
-public class AnnotationsExtendedMetaTypeService
-	implements ExtendedMetaTypeService {
+//@Component(service = ExtendedMetaTypeService.class)
+public class EquinoxExtendedMetaTypeService implements ExtendedMetaTypeService {
 
 	@Override
 	public ExtendedMetaTypeInformation getMetaTypeInformation(Bundle bundle) {
-		return new AnnotationsExtendedMetaTypeInformation(
-			bundle, _metaTypeService.getMetaTypeInformation(bundle));
+		return new EquinoxExtendedMetaTypeInformation(
+			_metaTypeService.getMetaTypeInformation(bundle));
 	}
 
 	@Reference(unbind = "-")
